Extract build command from CLI app
