+++ src/cli/commands/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(feature = "build")]
+pub(crate) mod build;
